test: CarNameValidator, TryCountValidator test 병합
